requirements.txt
git+https://github.com/SwanseaUniversityMedical/pyconceptlibraryclient.git@v1.0.0
beautifulsoup4==4.13.4
openai==1.75.0
pandas==2.2.3
pdfplumber==0.11.5
PyYAML==6.0.2
Requests==2.32.3
selenium==4.31.0
webdriver_manager==4.0.2
bs4
python-dotenv==0.21.0
neo4j==5.28.1
pyvis==0.3.1